Tri du code vers les dossiers niveaux.c/.h et appel des fonctions depuis Niveau1.c. Correction bug tapis roulant gauche qui ne marchait pas dans toutes les conditions.

+++ entitees/conversionCoordonnees.c
@@ -1,11 +1,11 @@
 #include "conversionCoordonnees.h"
 
 
-int convCoX(int x) {
+short convCoX(int x) {
     return ((x * 2) + MINX);
 }
 
-int convCoY(int y) {
+short convCoY(int y) {
     return ((y * 4) + MINY);
 }
