+++ entitees/conversionCoordonnees.h
@@ -8,8 +8,8 @@
 #define MINX 7
 #define MINY (MINX + 66)
 
-int convCoX(int x);
-int convCoY(int y);
+short convCoX(int x);
+short convCoY(int y);
 
 COORD convCo(int x, int y);
